Add ID generator
With a new script unique IDs (only for `xml:id`) of the form `X0000000` can be automated generated and it can be triggered by the corresponding action from Oxygen (ediarum.BASE). For every new project a ID namespace (one letter ist recommended) should be specified. The script `ediarum_idGenerator.xql` is copied to the project collection `/oxygen`. If this feature should be used in present projects `<id-namespace></id-namespace>` must be added manually to project `config.xml`.